+++ lib/clearance.rb
@@ -6,3 +6,9 @@ require 'clearance/user'
 require 'clearance/engine'
 require 'clearance/password_strategies'
 require 'clearance/constraints'
+
+module Clearance
+  def self.root
+    File.expand_path('../..', __FILE__)
+  end
+end